G-MSRA/requirements.txt
# G-MSRA Dependencies
# Install: pip install -r requirements.txt

# Core
torch>=2.1.0
transformers>=4.40.0
accelerate>=0.28.0
datasets>=2.18.0

# RL Training
trl>=0.8.0

# LoRA / Parameter-Efficient Fine-Tuning
peft>=0.10.0
bitsandbytes>=0.43.0  # QLoRA 4-bit support

# Retrieval
faiss-gpu>=1.7.4
sentence-transformers>=2.6.0

# Evaluation
rouge-score>=0.1.2
nltk>=3.8.1

# Utilities
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.12.0
scikit-learn>=1.4.0
pyyaml>=6.0.1
tqdm>=4.66.0
wandb>=0.16.0
loguru>=0.7.2
jsonlines>=4.0.0
